Emprestar o livro agora esta funcionando

+++ app/src/main/java/com/example/controlebiblioteca/classes/Emprestimo.kt
@@ -4,11 +4,11 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity(tableName = "emprestimo")
-data class Emprestimo (
+data class Emprestimo(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
     val dataEmprestimo: Long, //Para ser compatível com timestamps
     val dataDevolucao: Long?, //Pode ser nulo caso ainda não tenha sido devolvido
-    val livroId: Int,
+    val livroId: Long,
     val usuarioId: Int,
     val status: String
 )
